+++ src/test/converter2/behavior/declareGlobal.ts
@@ -1,0 +1,9 @@
+import { SyntaxKind } from "typescript";
+
+declare global {
+    interface DeclareGlobal {
+        method(kind: SyntaxKind): void;
+    }
+}
+
+namespace NotIncluded {}